Simplify the enum and move it to RenderContext
