+++ gapic/src/main/com/google/gapid/perfetto/models/Tracks.java
@@ -26,6 +26,7 @@ import com.google.common.collect.Lists;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.gapid.models.Perfetto;
 import com.google.gapid.perfetto.canvas.Panel;
+import com.google.gapid.perfetto.canvas.RenderContext.Truncate;
 import com.google.gapid.perfetto.models.TrackConfig.Group;
 import com.google.gapid.perfetto.views.CounterPanel;
 import com.google.gapid.perfetto.views.CpuFrequencyPanel;
@@ -36,7 +37,6 @@ import com.google.gapid.perfetto.views.GpuQueuePanel;
 import com.google.gapid.perfetto.views.ProcessSummaryPanel;
 import com.google.gapid.perfetto.views.ThreadPanel;
 import com.google.gapid.perfetto.views.TitlePanel;
-import com.google.gapid.perfetto.views.TrackContainer.Truncate;
 import com.google.gapid.perfetto.views.VulkanCounterPanel;
 import com.google.gapid.perfetto.views.VulkanEventPanel;
 import com.google.gapid.util.Scheduler;
